ML/model3_suggestions/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
openai>=1.68.2
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.26.0
